+++ scripts/large_file_test_script.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-# Copyright (c) 2016, 2020, Oracle and/or its affiliates.  All rights reserved.
+# Copyright (c) 2016, 2021, Oracle and/or its affiliates.  All rights reserved.
 # This software is dual-licensed to you under the Universal Permissive License (UPL) 1.0 as shown at https://oss.oracle.com/licenses/upl or Apache License 2.0 as shown at http://www.apache.org/licenses/LICENSE-2.0. You may choose either license.
 # Creates an object of the given size in MB, uploads it to Object Storage Service,
 # downloads it again, then cleans up. The time for each operation is recorded in seconds.