+++ services/bds/src/oci_cli_bds/bds_cli_extended.py
@@ -1,5 +1,5 @@
 # coding: utf-8
-# Copyright (c) 2016, 2020, Oracle and/or its affiliates.  All rights reserved.
+# Copyright (c) 2016, 2021, Oracle and/or its affiliates.  All rights reserved.
 # This software is dual-licensed to you under the Universal Permissive License (UPL) 1.0 as shown at https://oss.oracle.com/licenses/upl or Apache License 2.0 as shown at http://www.apache.org/licenses/LICENSE-2.0. You may choose either license.
 
 from oci_cli import cli_util
@@ -51,3 +51,26 @@ def bds_cloud_sql_group():
 bds_cli.bds_root_group.add_command(bds_cloud_sql_group)
 bds_cloud_sql_group.add_command(bds_cli.add_cloud_sql)
 bds_cloud_sql_group.add_command(bds_cli.remove_cloud_sql)
+
+
+# Add command: oci bds auto-scaling-config create | delete | get | edit | list
+@click.command('auto-scale-config', cls=CommandGroupWithAlias, help=bds_cli.add_auto_scaling_configuration.help)
+@cli_util.help_option_group
+def bds_auto_scaling_group():
+    pass
+
+
+bds_cli.bds_root_group.add_command(bds_auto_scaling_group)
+bds_auto_scaling_group.add_command(bds_cli.add_auto_scaling_configuration)
+cli_util.rename_command(bds_cli, bds_auto_scaling_group, bds_cli.add_auto_scaling_configuration, "create")
+bds_auto_scaling_group.add_command(bds_cli.remove_auto_scaling_configuration)
+cli_util.rename_command(bds_cli, bds_auto_scaling_group, bds_cli.remove_auto_scaling_configuration, "delete")
+bds_cli.bds_instance_group.commands.pop(bds_cli.get_auto_scaling_configuration.name)
+bds_auto_scaling_group.add_command(bds_cli.get_auto_scaling_configuration)
+cli_util.rename_command(bds_cli, bds_auto_scaling_group, bds_cli.get_auto_scaling_configuration, "get")
+bds_cli.bds_instance_group.commands.pop(bds_cli.update_auto_scaling_configuration.name)
+bds_auto_scaling_group.add_command(bds_cli.update_auto_scaling_configuration)
+cli_util.rename_command(bds_cli, bds_auto_scaling_group, bds_cli.update_auto_scaling_configuration, "edit")
+bds_cli.bds_instance_group.commands.pop(bds_cli.list_auto_scaling_configurations.name)
+bds_auto_scaling_group.add_command(bds_cli.list_auto_scaling_configurations)
+cli_util.rename_command(bds_cli, bds_auto_scaling_group, bds_cli.list_auto_scaling_configurations, "list")