+++ services/database/tests/unit/test_db_cli_extended.py
@@ -1,5 +1,5 @@
 # coding: utf-8
-# Copyright (c) 2016, 2020, Oracle and/or its affiliates.  All rights reserved.
+# Copyright (c) 2016, 2021, Oracle and/or its affiliates.  All rights reserved.
 # This software is dual-licensed to you under the Universal Permissive License (UPL) 1.0 as shown at https://oss.oracle.com/licenses/upl or Apache License 2.0 as shown at http://www.apache.org/licenses/LICENSE-2.0. You may choose either license.
 
 import unittest
@@ -237,6 +237,6 @@ class TestDBCliExtended(unittest.TestCase):
         assert '--vm-cluster-id' in result.output
 
     def test_list_db_flex_component(self):
-            result = util.invoke_command(['db', 'flex-component', 'list'])
-            assert 'Error: Missing option(s)' in result.output
-            assert '--compartment-id' in result.output
+        result = util.invoke_command(['db', 'flex-component', 'list'])
+        assert 'Error: Missing option(s)' in result.output
+        assert '--compartment-id' in result.output