+++ services/dts/src/oci_cli_appliance_export_job/manifest/manifest_writer.py
@@ -1,5 +1,5 @@
 # coding: utf-8
-# Copyright (c) 2016, 2020, Oracle and/or its affiliates.  All rights reserved.
+# Copyright (c) 2016, 2021, Oracle and/or its affiliates.  All rights reserved.
 # This software is dual-licensed to you under the Universal Permissive License (UPL) 1.0 as shown at https://oss.oracle.com/licenses/upl or Apache License 2.0 as shown at http://www.apache.org/licenses/LICENSE-2.0. You may choose either license.
 
 from .manifest_constants import ENCODING_FORMAT_UTF, MAX_EXPORT_SIZE, OBJECT_NAME_MAX_PATH_LENGTH, MAX_PART_SIZE_MB