+++ services/load_balancer/examples_and_test_scripts/create_listener_example.sh
@@ -1,7 +1,11 @@
 #!/bin/bash
+# Copyright (c) 2016, 2021, Oracle and/or its affiliates.  All rights reserved.
+# This software is dual-licensed to you under the Universal Permissive License (UPL) 1.0 as shown at https://oss.oracle.com/licenses/upl or Apache License 2.0 as shown at http://www.apache.org/licenses/LICENSE-2.0. You may choose either license.
+
 # -----------------------------------------
 # Create listener,backend_set,cipher_suite using individual ssl_configuration options only
 # -----------------------------------------
+
 
 # Requirements for running this script:
 # -----------------------------------------