+++ services/mysql/src/oci_cli_db_system/db_system_cli_extended.py
@@ -1,5 +1,9 @@
 # coding: utf-8
+# Copyright (c) 2016, 2021, Oracle and/or its affiliates.  All rights reserved.
+# This software is dual-licensed to you under the Universal Permissive License (UPL) 1.0 as shown at https://oss.oracle.com/licenses/upl or Apache License 2.0 as shown at http://www.apache.org/licenses/LICENSE-2.0. You may choose either license.
+
 # Copyright (c) 2016, 2019, Oracle and/or its affiliates. All rights reserved.
+
 
 from services.mysql.src.oci_cli_db_system.generated import dbsystem_cli
 from oci_cli import cli_util