+++ src/common_util/ignored_commands.py
@@ -1,3 +1,5 @@
+# Copyright (c) 2016, 2021, Oracle and/or its affiliates.  All rights reserved.
+# This software is dual-licensed to you under the Universal Permissive License (UPL) 1.0 as shown at https://oss.oracle.com/licenses/upl or Apache License 2.0 as shown at http://www.apache.org/licenses/LICENSE-2.0. You may choose either license.
 IGNORED_COMMANDS = [
     ['setup', 'autocomplete'],
     ['setup', 'bootstrap'],
@@ -81,4 +83,20 @@ IGNORED_COMMANDS = [
     ['network', 'private-endpoint', 'enable-reverse-connections'],
     ['os', 'object', 'reencrypt'],
     ['dns', 'resolver', 'update']
+]
+
+IGNORED_COMMANDS_DOCS = [
+    ['setup', 'autocomplete'],
+    ['setup', 'bootstrap'],
+    ['setup', 'config'],
+    ['setup', 'keys'],
+    ['setup', 'repair-file-permissions'],
+    ['setup', 'oci-cli-rc'],
+    ['raw-request'],
+    ['session', 'authenticate'],
+    ['session', 'export'],
+    ['session', 'import'],
+    ['session', 'refresh'],
+    ['session', 'terminate'],
+    ['session', 'validate'],
 ]